+++ src/js/themes/colors.js
@@ -1,0 +1,133 @@
+export const colors = {
+  /* deprecated accent and neutral colors */
+  'accent-1': undefined,
+  'accent-2': undefined,
+  'accent-3': undefined,
+  'accent-4': undefined,
+  'neutral-1': undefined,
+  'neutral-2': undefined,
+  'neutral-3': undefined,
+  'neutral-4': undefined,
+  'neutral-5': undefined,
+  'status-error': undefined,
+  brand: 'green!',
+  background: {
+    dark: '#263040',
+    light: '#FFFFFF',
+  },
+  'background-back': {
+    dark: '#263040',
+    light: '#EFEFEF',
+  },
+  'background-front': {
+    dark: '#404B5C',
+    light: '#FFFFFF',
+  },
+  'background-contrast': {
+    dark: '#FFFFFF14',
+    light: '#0000000A',
+  },
+  icon: 'text',
+  text: {
+    dark: '#C0CADC',
+    light: '#444444',
+  },
+  'text-strong': {
+    dark: '#FFFFFF',
+    light: '#000000',
+  },
+  'text-weak': {
+    dark: '#606B7D',
+    light: '#BBBBBB',
+  },
+  'text-xweak': 'text-weak',
+  border: {
+    dark: '#7887A1',
+    light: '#999999',
+  },
+  'border-strong': {
+    dark: '#AFBCD2',
+    light: '#666666',
+  },
+  'border-weak': {
+    dark: '#606B7D',
+    light: '#BBBBBB',
+  },
+  control: 'green',
+  'active-background': 'background-contrast',
+  'active-text': 'text',
+  'disabled-text': 'text-weak', // deprecated, use text-weak instead
+  'selected-background': 'green',
+  'selected-text': 'text-strong',
+  'status-critical': {
+    dark: '#D04F4E',
+    light: '#FC5A5A',
+  },
+  'status-warning': {
+    dark: '#9B6310',
+    light: '#FFBC44',
+  },
+  'status-ok': {
+    dark: '#008567',
+    light: '#17EBA0',
+  },
+  'status-unknown': {
+    dark: '#4F5F76',
+    light: '#CCCCCC',
+  },
+  'status-disabled': '#CCCCCC', // deprecated, does not support light and dark. use text-weak instead
+  blue: {
+    dark: '#00567A',
+    light: '#00C8FF',
+  },
+  'blue!': '#00739D',
+  green: {
+    dark: '#008567',
+    light: '#17EBA0',
+  },
+  'green!': '#01A982',
+  teal: {
+    dark: '#117B82',
+    light: '#82FFF2',
+  },
+  'teal!': '#00E8CF',
+  purple: {
+    dark: '#6633BC',
+    light: '#F740FF',
+  },
+  'purple!': '#7630EA',
+  red: {
+    dark: '#A2423D',
+    light: '#FC6161',
+  },
+  'red!': '#C54E4B',
+  orange: {
+    dark: '#9B6310',
+    light: '#FFBC44',
+  },
+  'orange!': '#FF8300',
+  yellow: {
+    dark: '#8D741C',
+    light: '#FFEB59',
+  },
+  'yellow!': '#FEC901',
+  'validation-critical': {
+    light: '#FC61613D',
+    dark: '#C54E4B5C',
+  },
+  'validation-ok': {
+    light: '#17EBA03D',
+    dark: '#00856759',
+  },
+  'validation-warning': {
+    light: '#FFBC443D',
+    dark: '#9B63105C',
+  },
+  'graph-0': 'orange!',
+  'graph-1': 'blue!',
+  'graph-2': 'purple!',
+  'graph-3': 'yellow!',
+  'graph-4': 'teal!',
+  focus: 'teal!',
+  placeholder: 'text-weak',
+};
