+++ src/js/themes/index.d.ts
@@ -1,0 +1,5 @@
+declare const hpe: object;
+declare const colors: object;
+declare const codeTheme: object;
+
+export { hpe };
