+++ src/js/themes/index.js
@@ -1,0 +1,3 @@
+export * from './code';
+export * from './colors';
+export * from './hpe';
